added logic to show title on load, event handler to change game state on game start, and refactored Pokemon API call to include the 3 pokemon that we need information about
